Conflicts:
	DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/config/HokenshaJohoConfig.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/ForeignersNameOutput.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/GappeiKyuShichosonHyoji.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/GappeiKyuShichosonKubun.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/GunNamePrint.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/HokenshaKosei.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/KatagakiPrint.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/RokenJukyushaNoTaikei.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/ShichosonNamePrint.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/TodofukenNamePrint.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/TopPriorityArea.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/enumeratedtype/hokensha/UnyoKeitaiKubun.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/valueobject/hokensha/GappeiChiikiNo.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/valueobject/hokensha/HokenshaName.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/valueobject/hokensha/KokuhorenKoikiShichosonNo.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/valueobject/hokensha/RokenShichosonNo.java
	DB/DBZ/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbz/definition/valueobject/hokensha/ShichosonShikibetsuID.java

+++ DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/basic/DbT2003KibetsuEntity.java
@@ -6,7 +6,6 @@ import jp.co.ndensan.reams.db.dbz.definition.valueobject.domain.TsuchishoNo;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RDateTime;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.DbTableEntityBase;
 import jp.co.ndensan.reams.uz.uza.util.db.IDbAccessable;
 import jp.co.ndensan.reams.uz.uza.util.db.PrimaryKey;
@@ -17,7 +16,7 @@ import jp.co.ndensan.reams.uz.uza.util.db.TableName;
  *
  */
 public class DbT2003KibetsuEntity extends DbTableEntityBase<DbT2003KibetsuEntity> implements IDbAccessable {
-// <editor-fold defaultstate="collapsed" desc="Created By POJO Tool ver 1.3.1">
+// <editor-fold defaultstate="collapsed" desc="Created By POJO Tool ver 1.3.7">
 
     @TableName
     public static final RString TABLE_NAME = new RString("DbT2003Kibetsu");
@@ -43,7 +42,7 @@ public class DbT2003KibetsuEntity extends DbTableEntityBase<DbT2003KibetsuEntity
     private RString choshuHoho;
     @PrimaryKey
     private int ki;
-    private Decimal choteiId;
+    private Long choteiId;
 
     /**
      * getInsertDantaiCd
@@ -203,7 +202,7 @@ public class DbT2003KibetsuEntity extends DbTableEntityBase<DbT2003KibetsuEntity
      *
      * @return choteiId
      */
-    public Decimal getChoteiId() {
+    public Long getChoteiId() {
         return choteiId;
     }
 
@@ -212,7 +211,7 @@ public class DbT2003KibetsuEntity extends DbTableEntityBase<DbT2003KibetsuEntity
      *
      * @param choteiId choteiId
      */
-    public void setChoteiId(Decimal choteiId) {
+    public void setChoteiId(Long choteiId) {
         this.choteiId = choteiId;
     }
